% Generated by roxygen2 (4.0.1): do not edit by hand
\name{callPayoffGH}
\alias{callPayoffGH}
\title{Payoff for call option settled by delivery: g=ST,h=-K if in the money}
\usage{
callPayoffGH(ST, K)
}
\arguments{
\item{ST}{stock price at option expiration}

\item{K}{call option strike price}
}
\description{
Payoff for call option settled by delivery: g=ST,h=-K if in the money
}

