% Generated by roxygen2 (4.0.1): do not edit by hand
\name{collectGHRutkowski}
\alias{collectGHRutkowski}
\title{Collect portfolio values, stock values, f1,f2 values, etc.}
\usage{
collectGHRutkowski(path, lambda = 0.2, mu = 1/6, r = 0, u = 1.3,
  d = 0.9, K = 100)
}
\arguments{
\item{path}{simulated stock price path}

\item{lambda}{purchase cost parameter}

\item{mu}{sale cost parameter}

\item{r}{risk-free interest rate}

\item{u}{u parameter}

\item{d}{d parameter}

\item{K}{option strike price}
}
\description{
Collect portfolio values, stock values, f1,f2 values, etc.
}
\note{
All but path have default values.
}

