% Generated by roxygen2 (4.0.1): do not edit by hand
\name{computeF1F2}
\alias{computeF1F2}
\title{fold back f1f2 tree}
\usage{
computeF1F2(stockList, S0, u, d, nFlips, K, lambda, mu, phat, ptilda, r)
}
\arguments{
\item{stockList}{List of stock tree node values}

\item{S0}{initial stock price}

\item{u}{u parameter}

\item{d}{d parameter}

\item{nFlips}{number of rebalances}

\item{K}{option strike price}

\item{lambda}{purchase cost parameter}

\item{mu}{sale cost parameter}

\item{phat}{probability phat}

\item{ptilda}{probability ptilda}

\item{r}{risk-free interest rate}
}
\description{
fold back f1f2 tree
}

