% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createDeltaRutkowski}
\alias{createDeltaRutkowski}
\title{loops over stock price paths and computes delta, G, H
del is delta with NA appended at time 0. Seems to not be used.}
\usage{
createDeltaRutkowski(pathMatrix, lambda, mu, r, u, d, K)
}
\arguments{
\item{pathMatrix}{matrix of simulated stock price paths}

\item{lambda}{purchase cost parameter}

\item{mu}{sale cost parameter}

\item{r}{risk-free interest rate}

\item{u}{u parameter}

\item{d}{d parameter}

\item{K}{option strike prices}
}
\description{
loops over stock price paths and computes delta, G, H
del is delta with NA appended at time 0. Seems to not be used.
}

