% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createDeltaRutkowskiNew}
\alias{createDeltaRutkowskiNew}
\title{Computes small delta and portfolio G,H given a (u,d) pair}
\usage{
createDeltaRutkowskiNew(u, d)
}
\arguments{
\item{u}{u parameter}

\item{d}{d parameter}
}
\value{
outputList A list consisting of delta, GMatrix, HMatrix, pathMatrix
}
\description{
Calls collectGHRutkowski() and rutkowskiDelta() and returns a list
consisting of delta, portfolio matrices GMatrix, HMatrix and
matrix pathMatrix, all for one (u,d) pair.
}

