% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createTCosts}
\alias{createTCosts}
\title{Computes time series of costs when share component of portfolio changes}
\usage{
createTCosts(sPath, portShareQt, lambda, mu, verbose = FALSE)
}
\arguments{
\item{sPath}{a simulated stock price path}

\item{portShareQt}{quantity of shares in portfolio}

\item{lambda}{purchase cost parameter}

\item{mu}{sale cost parameter}

\item{verbose}{if TRUE, print some information; default is FALSE}
}
\description{
Computes time series of costs when share component of portfolio changes
}

