% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fAtExpiration}
\alias{fAtExpiration}
\title{Expiration value of functions f1 and f2}
\usage{
fAtExpiration(g, h, lambda, mu)
}
\arguments{
\item{g}{stock value component of portfolio}

\item{h}{bond value component of portfolio}

\item{lambda}{purchase cost parameter}

\item{mu}{sale cost parameter}
}
\description{
Expiration value of functions f1 and f2
}

