% Generated by roxygen2 (4.0.1): do not edit by hand
\name{HG}
\alias{HG}
\title{Uses functions f1, f2 to compute portfolio components H,G}
\usage{
HG(mu, lambda, fStored)
}
\arguments{
\item{mu}{sale cost parameter}

\item{lambda}{purchase cost parameter}

\item{fStored}{list of stored values of functions f1 and f2}
}
\description{
Uses functions f1, f2 to compute portfolio components H,G
}

