% Generated by roxygen2 (4.0.1): do not edit by hand
\name{optionPriceRut}
\alias{optionPriceRut}
\title{Folded back tree of f values to compute option price at purchase}
\usage{
optionPriceRut(phat, r, fStored)
}
\arguments{
\item{phat}{probability phat}

\item{r}{risk-free interest rate}

\item{fStored}{list of stored f1,f2 values}
}
\description{
Folded back tree of f values to compute option price at purchase
}

