% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prob}
\alias{prob}
\title{Computes probabilities phat and ptilda}
\usage{
prob(lambda, mu, r, u, d)
}
\arguments{
\item{lambda}{purchase cost parameter}

\item{mu}{sale cost parameter}

\item{r}{risk-free interest rate}

\item{u}{u parameter}

\item{d}{d parameter}
}
\description{
Computes probabilities phat and ptilda
}

