% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rutkowskiDelta}
\alias{rutkowskiDelta}
\title{Compute delta time series using Rutkowski portfolio G,H}
\usage{
rutkowskiDelta(path, H, G, lambda = 0.2, mu = 1/6, r = 0)
}
\arguments{
\item{path}{stock price path}

\item{H}{portfolio bond value H}

\item{G}{portfolio stock value G}

\item{lambda}{purchase cost parameter}

\item{mu}{sale cost parameter}

\item{r}{risk-free interest rate}
}
\description{
Compute delta time series using Rutkowski portfolio G,H
}
\note{
lambda has default value 0.2
mu has default value 1/6
r has default value 0.
}

