% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stock}
\alias{stock}
\title{Computes list of values S0*u^j*d^(nFlips - j)}
\usage{
stock(S0, u, d, nFlips)
}
\arguments{
\item{S0}{Initial stock price}

\item{u}{u parameter}

\item{d}{d parameter}

\item{nFlips}{Number of rebalances}
}
\description{
Computes list of values S0*u^j*d^(nFlips - j)
}

