% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tCostRutkowski}
\alias{tCostRutkowski}
\title{constructs transaction cost using Rutkowski model}
\usage{
tCostRutkowski(pathMatrix, GMatrix, lambda, mu)
}
\arguments{
\item{pathMatrix}{matrix of simulated stock price paths}

\item{GMatrix}{matrix of stock component of portfolio}

\item{lambda}{purchase cost parameter}

\item{mu}{sale cost parameter}
}
\description{
constructs transaction cost using Rutkowski model
}
\note{
This function is not used in the simulation
}

