% Generated by roxygen2 (4.0.1): do not edit by hand
\name{timeNames}
\alias{timeNames}
\title{Utility functions for calculating delta and portfolio G,H matrices}
\usage{
timeNames(nTimes)
}
\arguments{
\item{nTimes}{Number of values in time series.}
}
\description{
This file contains various functions needed to compute delta and portfolios.
}
\details{
Constructs time names of the form time=number for labeling rows of matrices
}

