% Generated by roxygen2 (4.0.1): do not edit by hand
\name{updateF}
\alias{updateF}
\title{Update functions f1 and f2}
\usage{
updateF(phat, ptilda, r, jTime, f1, f2)
}
\arguments{
\item{phat}{Probability phat}

\item{ptilda}{probability ptilda}

\item{r}{risk-free interest rate}

\item{jTime}{current time}

\item{f1}{current value of f1}

\item{f2}{current value of f2}
}
\description{
Update functions f1 and f2
}

